moved ResponseCodes.java into a more suitable package
