+++ src/main/java/de/bywambo/chattodiscord/utils/Connector.java
@@ -1,6 +1,6 @@
 package de.bywambo.chattodiscord.utils;
 
-import de.bywambo.chattodiscord.models.ResponseCodes;
+import de.bywambo.chattodiscord.constants.ResponseCodes;
 import org.bukkit.Bukkit;
 
 import java.io.IOException;
